# openapi.yml
openapi: 3.0.3
info:
  title: EE3070 Project Server API
  version: 1.0.0
  description: |
    OpenAPI 規格依據目前 `src/routes` 的實際實作整理。
    主要功能包含課程、檔案上傳、題目產生任務、取消與狀態串流 (SSE)。
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /course:
    get:
      tags: [ Courses ]
      summary: List courses
      operationId: listCourses
      responses:
        '200':
          description: Course list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
              example:
                code: 200
                message: Course list retrieved successfully
                data:
                  - courseId: 546644f3-9632-42b1-a916-79c1657f9938
                    name: Test Course2
                    digitId: "000001"
                  - courseId: 9b58e4d1-4ae3-4f60-94bd-04e63fc6be44
                    name: Test Course1
                    digitId: "000002"
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [ Courses ]
      summary: Create course
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
            example:
              name: Introduction to Biology
      responses:
        '200':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseCreateResponse'
              example:
                code: 200
                message: Course created successfully
                data:
                  courseId: 9b58e4d1-4ae3-4f60-94bd-04e63fc6be44
                  digitId: "000002"
                  name: Test Course1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /course/{courseId}:
    get:
      tags: [ Courses ]
      summary: Get course info
      operationId: getCourse
      parameters:
        - $ref: '#/components/parameters/CourseIdPath'
      responses:
        '200':
          description: Course info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseInfoResponse'
              example:
                code: 200
                message: Course info retrieved successfully
                data:
                  courseId: 546644f3-9632-42b1-a916-79c1657f9938
                  name: Test Course2
                  digitId: "000001"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    CourseIdPath:
      name: courseId
      in: path
      required: true
      description: Course UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: Invalid request
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            message: Resource not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: Internal server error
  schemas:
    ErrorResponse:
      type: object
      required: [ code, message ]
      properties:
        code:
          type: integer
        message:
          type: string
...
    CourseItem:
      type: object
      required: [ name, digitId ]
      properties:
        id:
          type: string
          format: uuid
          description: Legacy field (may still appear in old cached records)
        courseId:
          type: string
          format: uuid
        name:
          type: string
        digitId:
          type: string
          pattern: '^\d{6}$'
    CourseCreateRequest:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
          minLength: 1
    CourseListResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Course list retrieved successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseItem'
    CourseCreateResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Course created successfully
        data:
          type: object
          required: [ courseId, digitId, name ]
          properties:
            courseId:
              type: string
              format: uuid
            digitId:
              type: string
              pattern: '^\d{6}$'
            name:
              type: string
    CourseInfoResponse:
      type: object
      required: [ code, message, data ]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Course info retrieved successfully
        data:
          $ref: '#/components/schemas/CourseItem'
...